fix(arch): satisfy size ratchets after rebase onto main
The rebase landed on a main where four touched files sit at their arch
size limits. Keep every scan-pinned construct in place and restore the
limits by relocation/condensation only:

- query_boundaries/assignability.rs: re-join the import groups (the
  'let analysis = ...' split is scan-pinned and stays).
- assignability_checker.rs: move the self-contained 'typeof this'
  guard-narrowing cluster to assignability/typeof_this_guard.rs; the
  scan-pinned ensure_relation_pair_ready / is_relation_cacheable gating
  stays in place.
- error_reporter/properties.rs: route format_excess_property_name
  through the existing 'query' module alias (one line).
- state/state_checking/property.rs: local module alias for the
  boundary call (one line).

No behavior change; arch guard passes; contract suite 138/142 with the
same 4 pre-existing failures documented in the PR body.
