+++ crates/tsz-checker/src/assignability/mod.rs
@@ -22,3 +22,4 @@ mod polymorphic_this_diagnostics;
 mod readonly_tuple_diagnostics;
 mod relation_outcome_helpers;
 pub mod subtype_identity_checker;
+mod typeof_this_guard;
