+++ crates/tsz-checker/src/assignability/typeof_this_guard.rs
@@ -1,0 +1,130 @@
+//! `typeof this === "..."` guard narrowing for assignability call sites.
+//!
+//! When an assignability diagnostic is about to be reported for an expression
+//! involving `this`, an enclosing `typeof this === "function"`-style guard can
+//! narrow `this` first. These helpers locate that guard and produce the
+//! narrowed type so diagnostics reflect the guarded type, mirroring tsc.
+
+use crate::query_boundaries::common::NarrowingContext;
+use crate::state::CheckerState;
+use tsz_parser::parser::NodeIndex;
+use tsz_parser::parser::syntax_kind_ext;
+use tsz_scanner::SyntaxKind;
+use tsz_solver::TypeId;
+
+impl<'a> CheckerState<'a> {
+    fn typeof_this_comparison_literal(
+        &self,
+        left: NodeIndex,
+        right: NodeIndex,
+        this_ref: NodeIndex,
+    ) -> Option<&str> {
+        if self.is_typeof_this_target(left, this_ref) {
+            return self.string_literal_text(right);
+        }
+        if self.is_typeof_this_target(right, this_ref) {
+            return self.string_literal_text(left);
+        }
+        None
+    }
+
+    fn is_typeof_this_target(&self, expr: NodeIndex, this_ref: NodeIndex) -> bool {
+        let expr = self.ctx.arena.skip_parenthesized(expr);
+        let Some(node) = self.ctx.arena.get(expr) else {
+            return false;
+        };
+        if node.kind != syntax_kind_ext::PREFIX_UNARY_EXPRESSION {
+            return false;
+        }
+        let Some(unary) = self.ctx.arena.get_unary_expr(node) else {
+            return false;
+        };
+        if unary.operator != SyntaxKind::TypeOfKeyword as u16 {
+            return false;
+        }
+        let operand = self.ctx.arena.skip_parenthesized(unary.operand);
+        if operand == this_ref {
+            return true;
+        }
+        self.ctx
+            .arena
+            .get(operand)
+            .is_some_and(|n| n.kind == SyntaxKind::ThisKeyword as u16)
+    }
+
+    fn string_literal_text(&self, idx: NodeIndex) -> Option<&str> {
+        let idx = self.ctx.arena.skip_parenthesized(idx);
+        let node = self.ctx.arena.get(idx)?;
+        if node.kind == SyntaxKind::StringLiteral as u16
+            || node.kind == SyntaxKind::NoSubstitutionTemplateLiteral as u16
+        {
+            return self
+                .ctx
+                .arena
+                .get_literal(node)
+                .map(|lit| lit.text.as_str());
+        }
+        None
+    }
+
+    pub(crate) fn narrow_this_from_enclosing_typeof_guard(
+        &self,
+        source_idx: NodeIndex,
+        source: TypeId,
+    ) -> TypeId {
+        let is_this_source = self
+            .ctx
+            .arena
+            .get(source_idx)
+            .is_some_and(|n| n.kind == SyntaxKind::ThisKeyword as u16);
+        if !is_this_source {
+            return source;
+        }
+
+        let mut current = source_idx;
+        let mut depth = 0usize;
+        while depth < 256 {
+            depth += 1;
+            let Some(ext) = self.ctx.arena.get_extended(current) else {
+                break;
+            };
+            if ext.parent.is_none() {
+                break;
+            }
+            current = ext.parent;
+            let Some(parent_node) = self.ctx.arena.get(current) else {
+                break;
+            };
+            if parent_node.kind != syntax_kind_ext::IF_STATEMENT {
+                continue;
+            }
+            let Some(if_stmt) = self.ctx.arena.get_if_statement(parent_node) else {
+                continue;
+            };
+            if !self.is_node_within(source_idx, if_stmt.then_statement) {
+                continue;
+            }
+            let Some(cond_node) = self.ctx.arena.get(if_stmt.expression) else {
+                continue;
+            };
+            if cond_node.kind != syntax_kind_ext::BINARY_EXPRESSION {
+                continue;
+            }
+            let Some(bin) = self.ctx.arena.get_binary_expr(cond_node) else {
+                continue;
+            };
+            let is_equality = bin.operator_token == SyntaxKind::EqualsEqualsEqualsToken as u16
+                || bin.operator_token == SyntaxKind::EqualsEqualsToken as u16;
+            if !is_equality {
+                continue;
+            }
+            if let Some(type_name) =
+                self.typeof_this_comparison_literal(bin.left, bin.right, source_idx)
+            {
+                return NarrowingContext::new(self.ctx.types).narrow_by_typeof(source, type_name);
+            }
+        }
+
+        source
+    }
+}
