+++ crates/tsz-checker/src/error_reporter/properties.rs
@@ -437,9 +437,7 @@ impl<'a> CheckerState<'a> {
     ) -> (u32, String) {
         let prop_display = self
             .excess_property_name_display_for_site(idx, self.ctx.types.intern_string(prop_name))
-            .unwrap_or_else(|| {
-                crate::query_boundaries::common::format_excess_property_name(prop_name).into_owned()
-            });
+            .unwrap_or_else(|| query::format_excess_property_name(prop_name).into_owned());
         let type_str = self.excess_property_target_display_for_site(target, idx);
         let suggestion_target = self.strip_non_object_union_members_for_excess_display(target);
         if !self.has_syntax_parse_errors()
