+++ crates/tsz-checker/src/state/state_checking/property.rs
@@ -544,10 +544,9 @@ impl<'a> CheckerState<'a> {
                 // property checking when it appears in a union.  Detect it by
                 // checking whether ALL property names are standard Object.prototype
                 // methods.  Similarly, skip for `Function` (has bind, call, apply, etc.).
-                if crate::query_boundaries::assignability::is_global_object_or_function_shape_boundary(
-                    self.ctx.types,
-                    &shape,
-                ) {
+                use crate::query_boundaries::assignability as assign_query;
+                if assign_query::is_global_object_or_function_shape_boundary(self.ctx.types, &shape)
+                {
                     return;
                 }
